Disable relational MiniPlaceholders by default
With MP v2, we would add "non-relational placeholders for sender (principal)" and then "placeholders for viewer (principal), in relation to sender", as was our understanding of how relational placeholders should generally be used in PlaceholderAPI and MiniPlaceholders.

With MP v3, we must add both tag resolvers and pass a relational audience to MiniMessage. The issue is that the principal audience for non-relational placeholders then becomes the viewer instead of the sender.

Pending any change in how relational MiniPlaceholders work (or how they should be used according to the dev), we will disable them by default to reduce the need for render_for tags to effectively disable this new principal audience swap behavior in formats.

If relational support is desired, you will need to deal with the consequences, including how preprocess tags inherently do not play nice with render_for due to how MiniMessage parsing works.